+++ src/Views/Shared/Master.Master.designer.cs
@@ -49,15 +49,6 @@ namespace lovis.Views.Shared {
         protected global::System.Web.UI.HtmlControls.HtmlAnchor Master_pagetitle;
         
         /// <summary>
-        /// Control btnSearch.
-        /// </summary>
-        /// <remarks>
-        /// Campo generado automáticamente.
-        /// Para modificarlo, mueva la declaración del campo del archivo del diseñador al archivo de código subyacente.
-        /// </remarks>
-        protected global::System.Web.UI.HtmlControls.HtmlButton btnSearch;
-        
-        /// <summary>
         /// Control master_username.
         /// </summary>
         /// <remarks>